moving apply filter to ast
